Refactor: TechStackSelectInput에 api요청추가
options를 받아오는 로직
